genie_python/requirements.txt
requests>=2.31.0
openai
fastapi
uvicorn
python-dotenv